Lots of changes
Went a little crazy, created art for the tanks, created a player class,
added recolor to entity class. Split the entity class in to 2 classes,
on that handles the entity pool, ents, and then the actual entity class.
Also started a view Information class that will track the view,
location, zoom, etc.